+++ src/TrashLib/Services/CustomFormat/Processors/GuideSteps/IQualityProfileStep.cs
@@ -6,5 +6,6 @@ public interface IQualityProfileStep
 {
     IDictionary<string, QualityProfileCustomFormatScoreMapping> ProfileScores { get; }
     IReadOnlyCollection<(string name, string trashId, string profileName)> CustomFormatsWithoutScore { get; }
+    IReadOnlyDictionary<string, Dictionary<string, HashSet<int>>> DuplicateScores { get; }
     void Process(IEnumerable<ProcessedConfigData> configData);
 }